Develop

+++ backend/src/chat/chat.gateway.ts
@@ -25,6 +25,7 @@ import { ChatSettingDto } from "./dto/chat-setting.dto";
 import { ChatAuthorityDto } from "./dto/chat-authority.dto";
 import { Block } from "src/block/block.entity";
 import { BlockRepository } from "src/block/block.repository";
+import { DMRepository } from "./repository/dm.repository";
 
 @WebSocketGateway({
   cors: {
@@ -37,49 +38,47 @@ export class ChatGateway {
     private chatParticipantsRepository: ChatParticipantRepository,
     private userRepository: UserRepository,
     private blockRepository: BlockRepository,
+    private dmRepository: DMRepository,
   ) {}
 
   @WebSocketServer()
   server;
-
-  // //////////////////////
-  // @SubscribeMessage("chat-room-all")
-  // async handleChatRoomAll(@ConnectedSocket() socket: Socket): Promise<void> {
-  //   console.log("chat room all");
-  //   const chatRooms: ChatRoom[] = await this.chatRoomRepository.find({
-  //     order: { id: 1 },
-  //   });
-
-  //   const chatRoomDto: ChatRoomDto[] = [];
-
-  //   chatRooms.forEach((chatRoom) => {
-  //     chatRoomDto.push(new ChatRoomDto(chatRoom));
-  //   });
-  //   this.server.emit("chat-room-all", chatRoomDto);
-  // }
 
   @SubscribeMessage("chat-room-join")
   async handleJoinChatRoom(
     @ConnectedSocket() socket: Socket,
     @MessageBody() chatJoinDto: ChatJoinDto,
   ): Promise<void> {
-    console.log("chat-room-join");
     const user: User = await this.userRepository.findOne(chatJoinDto.userId);
     chatJoinDto.nickname = user.nickname;
     let isCreate: Boolean = false;
 
-    // 방이 생성이면 생성 된 room id, 입장이면 프론트에서 넘어온 room id
     if (!chatJoinDto.chatRoomId) {
       chatJoinDto.chatRoomId = await this.chatRoomRepository.createRoom(
         chatJoinDto,
+        false,
       );
       isCreate = true;
     }
 
-    // 비번 틀리면 return
-    if (!(await this.joinChatRoom(chatJoinDto, user, isCreate))) {
+    const chatRoom: ChatRoom = await this.chatRoomRepository.findOne(
+      chatJoinDto.chatRoomId,
+    );
+    if (!chatRoom) return;
+
+    const joinUser: ChatParticipant =
+      await this.chatParticipantsRepository.findOne({
+        where: {
+          chatRoom: chatJoinDto.chatRoomId,
+          user: user.id,
+        },
+      });
+    if (joinUser) {
+      socket.join("chat-" + chatJoinDto.chatRoomId);
       return;
     }
+
+    if (!(await this.joinChatRoom(chatJoinDto, user, isCreate))) return;
 
     const chatRoomDto: ChatRoomDto = {
       chatRoomId: chatJoinDto.chatRoomId,
@@ -97,13 +96,11 @@ export class ChatGateway {
       .emit("chat-room-join", chatJoinDto);
   }
 
-  // setting
   @SubscribeMessage("chat-room-setting")
   async handleRoomSetting(
     @ConnectedSocket() socket: Socket,
     @MessageBody() chatSettingDto: ChatSettingDto,
   ) {
-    console.log("chat-room-setting");
     const chatRoom: ChatRoom = await this.chatRoomRepository.findOne(
       chatSettingDto.chatRoomId,
     );
@@ -119,13 +116,11 @@ export class ChatGateway {
 
     await chatRoom.save();
 
-    // await this.handleChatRoomAll(socket);
-    this.server
-      // .in("chat-" + chatSettingDto.chatRoomId)
-      .emit("chat-room-setting", {
-        chatRoomId: chatSettingDto.chatRoomId,
-        title: chatSettingDto.title,
-      });
+    this.server.emit("chat-room-setting", {
+      chatRoomId: chatSettingDto.chatRoomId,
+      title: chatSettingDto.title,
+      isPassword: chatRoom.password ? true : false,
+    });
   }
 
   @SubscribeMessage("chat-room-message")
@@ -134,7 +129,7 @@ export class ChatGateway {
     @MessageBody() chatMessageDto: ChatMessageDto,
   ): Promise<void> {
     const user: User = await this.userRepository.findOne(chatMessageDto.userId);
-    //채팅 금지 확인
+    chatMessageDto.nickname = user.nickname;
 
     const chatParticipant: ChatParticipant =
       await this.chatParticipantsRepository.findOne({ where: { user } });
@@ -147,15 +142,23 @@ export class ChatGateway {
       });
       return;
     }
+
+    const chatRoom: ChatRoom = await this.chatRoomRepository.findOne(
+      chatMessageDto.chatRoomId,
+    );
+
     const blocks: Block[] = await this.blockRepository.find({
       where: { blockId: user.id },
       relations: ["ownId"],
     });
+
     let temp = this.server.in("chat-" + chatMessageDto.chatRoomId);
+
+    if (chatRoom.isDM)
+      await this.dmRepository.createDM(chatRoom, user, chatMessageDto);
     blocks.forEach((block) => {
       temp = temp.except(block.ownId.socketId);
     });
-    chatMessageDto.nickname = user.nickname;
     temp.emit("chat-room-message", chatMessageDto);
   }
 
@@ -164,7 +167,6 @@ export class ChatGateway {
     @ConnectedSocket() socket: Socket,
     @MessageBody() setAdminDto: SetAdminDto,
   ): Promise<void> {
-    console.log("chat-room-set-admin");
     const chatParticipant: ChatParticipant =
       await this.chatParticipantsRepository.findOne({
         where: {
@@ -202,8 +204,6 @@ export class ChatGateway {
     this.server.in(target.socketId).socketsLeave("chat-" + chatRoomId);
   }
 
-  ///////////////////////////
-
   @SubscribeMessage("chat-room-leave")
   async handleChatRoomLeave(
     @ConnectedSocket() socket: Socket,
@@ -214,9 +214,6 @@ export class ChatGateway {
 
     const chatTitle = "chat-" + chatLeaveDto.chatRoomId;
 
-    // socket.leave("chat-" + chatLeaveDto.chatRoomId); //방을 떠난다.
-
-    // 참여자 리스트에서 그 룸의 그 참여자 삭제!
     const delUser: ChatParticipant =
       await this.chatParticipantsRepository.findOne({
         where: {
@@ -224,53 +221,27 @@ export class ChatGateway {
           user: user.id,
         },
       });
-    // const userAuthority = delUser.authority;
-    await this.chatParticipantsRepository.delete(delUser);
+    if (delUser) await this.chatParticipantsRepository.delete(delUser);
+    else return;
 
-    // this.chatParticipantAll(chatLeaveDto.chatRoomId);
     this.server.in(chatTitle).emit("chat-room-leave", chatLeaveDto);
-    socket.leave(chatTitle); //방을 떠난다.
+    socket.leave(chatTitle);
 
     const participant: ChatParticipant =
       await this.chatParticipantsRepository.findOne({
         where: { chatRoom: chatLeaveDto.chatRoomId },
       });
-    // if (!participant || userAuthority == Authority.owner) {
+
     if (!participant) {
-      this.server.emit("chat-room-destroy", {
-        chatRoomId: chatLeaveDto.chatRoomId,
-      });
       await this.chatParticipantsRepository.deleteAllParticipants(
         chatLeaveDto.chatRoomId,
       );
       await this.chatRoomRepository.deleteRoom(chatLeaveDto.chatRoomId);
-      // this.handleDisconnectChatRoom(chatLeaveDto.chatRoomId);
+      this.server.emit("chat-room-destroy", {
+        chatRoomId: chatLeaveDto.chatRoomId,
+      });
     }
   }
-
-  // @SubscribeMessage("connect-chat-room")
-  // handleConnectChatRoom(@MessageBody() chatJoinDto: ChatJoinDto) {
-  //   const chatTitle = "chat-" + chatJoinDto.chatRoomId;
-  //   const numParticipant: number =
-  //     this.server.sockets.adapter.rooms.get(chatTitle).size;
-  //   if (!numParticipant)
-  //     throw new NotFoundException(
-  //       `Can't find ChatRoom with ${chatJoinDto.title}`,
-  //     );
-  //   const chatRoomDto: ChatRoomDto = {
-  //     chatRoomId: chatJoinDto.chatRoomId,
-  //     title: chatJoinDto.title,
-  //     // password: chatJoinDto.password,
-  //     password: true,
-  //     // numParticipant,
-  //   };
-  //   this.server.emit("connect-chat-room", chatRoomDto);
-  // }
-
-  // @SubscribeMessage("disconnect-chat-room")
-  // handleDisconnectChatRoom(@MessageBody() chatRoomId: number) {
-  //   this.server.emit("disconnect-chat-room", chatRoomId); //? ㅇㅣ렇게 줘도 괜찮나?
-  // }
 
   private async joinChatRoom(
     chatJoinDto: ChatJoinDto,
@@ -281,16 +252,6 @@ export class ChatGateway {
       chatJoinDto.chatRoomId,
     );
 
-    const joinUser: ChatParticipant =
-      await this.chatParticipantsRepository.findOne({
-        where: {
-          chatRoom: chatJoinDto.chatRoomId,
-          user: chatJoinDto.userId,
-        },
-      });
-
-    if (joinUser) return true;
-
     const chatParticipants: ChatParticipant =
       this.chatParticipantsRepository.create({
         user,
@@ -299,18 +260,16 @@ export class ChatGateway {
 
     if (isCreate) chatParticipants.authority = Authority.OWNER;
 
-    if (chatJoinDto.password) {
-      if (await bcrypt.compare(chatJoinDto.password, chatRoom.password)) {
-        return true;
-      } else {
-        return false;
-      }
+    if (
+      chatJoinDto.password &&
+      !(await bcrypt.compare(chatJoinDto.password, chatRoom.password))
+    ) {
+      return false;
     }
 
     await chatParticipants.save();
     chatJoinDto.authority = chatParticipants.authority;
 
-    // this.handleConnectChatRoom(chatJoinDto);
     return true;
   }
 
@@ -331,13 +290,11 @@ export class ChatGateway {
     participants.forEach((participant) => {
       chatParticipantDto.push(new ChatParticipantDto(participant));
     });
-    // this.server.in("chat-" + roomId, chatParticipantDto);
     this.server
       .in("chat-" + roomId)
       .emit("chat-particip-all", chatParticipantDto);
   }
 
-  //채팅금지 30초
   @SubscribeMessage("chat-block")
   async handleChatBlock(
     @ConnectedSocket() socket: Socket,
@@ -348,7 +305,48 @@ export class ChatGateway {
         where: { user: targetId },
       });
     target.chatBlock = String(new Date().getTime());
-    // console.log(new Date().getTime());
     target.save();
+  }
+
+  @SubscribeMessage("DM")
+  async handleDM(
+    @ConnectedSocket() socket: Socket,
+    @MessageBody() { senderId, targetId },
+  ) {
+    const sender: User = await this.userRepository.findOne(senderId);
+    const target: User = await this.userRepository.findOne(targetId);
+    const dmTitle: string =
+      "dm- " +
+      Math.min(targetId, senderId) +
+      " + " +
+      Math.max(targetId, senderId);
+
+    let chatRoom: ChatRoom = await this.chatRoomRepository.findOne({
+      where: { title: dmTitle },
+    });
+
+    let isCreate: Boolean = false;
+
+    const chatJoinDto: ChatJoinDto = {
+      chatRoomId: 0,
+      title: dmTitle,
+      password: "",
+      userId: sender.id,
+      nickname: sender.nickname,
+      authority: Authority.PARTICIPANT,
+    };
+    if (!chatRoom) {
+      chatJoinDto.chatRoomId = await this.chatRoomRepository.createRoom(
+        chatJoinDto,
+        true,
+      );
+      isCreate = true;
+      await this.joinChatRoom(chatJoinDto, sender, isCreate);
+      await this.joinChatRoom(chatJoinDto, target, isCreate);
+      chatRoom = await this.chatRoomRepository.findOne(chatJoinDto.chatRoomId);
+    }
+    socket.join("chat-" + chatRoom.id);
+    this.server.in(target.socketId).socketsJoin("chat-" + chatRoom.id);
+    socket.emit("DM", { chatRoomId: chatRoom.id, targetId });
   }
 }
